+++ src/chrome/scripts/skipto.js
@@ -224,7 +224,7 @@
 
   const styleTemplate = document.createElement('template');
   styleTemplate.innerHTML = `
-<style type="text/css" id="id-skip-to-css">
+<style type="text/css">
 $skipToId.popup {
   top: -30px;
   transition: top 0.35s ease;
@@ -517,7 +517,12 @@ $skipToId [role="menuitem"].hover .label {
   background-color: $menuitemFocusBackgroundColor;
   color: $menuitemFocusTextColor;
 }
+</style>
+`;
 
+  const highlightTemplate = document.createElement('template');
+  highlightTemplate.innerHTML = `
+<style type="text/css">
 $skipToId-highlight {
   position: absolute;
   border-radius: 3px;
@@ -532,7 +537,6 @@ $skipToId-highlight div {
   border: 2px solid $focusBorderColor;
   z-index: $zHighlight;
 }
-
 </style>
 `;
 
@@ -625,7 +629,7 @@ $skipToId-highlight div {
    *
    *   @returns 
    */
-  function updateStyle(stylePlaceholder, configValue, themeValue, defaultValue) {
+  function updateStyle(template, stylePlaceholder, configValue, themeValue, defaultValue) {
     let value = defaultValue;
     if (typeof configValue === 'string' && configValue) {
       value = configValue;
@@ -635,7 +639,7 @@ $skipToId-highlight div {
       }
     }
 
-    let cssContent = styleTemplate.innerHTML;
+    let cssContent = template.innerHTML;
     let index1 = cssContent.indexOf(stylePlaceholder);
     let index2 = index1 + stylePlaceholder.length;
     while (index1 >= 0 && index2 < cssContent.length) {
@@ -643,7 +647,7 @@ $skipToId-highlight div {
       index1 = cssContent.indexOf(stylePlaceholder, index2);
       index2 = index1 + stylePlaceholder.length;
     }
-    styleTemplate.innerHTML = cssContent;
+    template.innerHTML = cssContent;
   }
 
   /*
@@ -664,27 +668,30 @@ $skipToId-highlight div {
       config.displayOption = theme.displayOption;
     }
 
-    updateStyle('$fontFamily', config.fontFamily, theme.fontFamily, defaultTheme.fontFamily);
-    updateStyle('$fontSize', config.fontSize, theme.fontSize, defaultTheme.fontSize);
+    updateStyle(styleTemplate, '$fontFamily', config.fontFamily, theme.fontFamily, defaultTheme.fontFamily);
+    updateStyle(styleTemplate, '$fontSize', config.fontSize, theme.fontSize, defaultTheme.fontSize);
 
-    updateStyle('$positionLeft', config.positionLeft, theme.positionLeft, defaultTheme.positionLeft);
-    updateStyle('$smallBreakPoint', config.smallBreakPoint, theme.smallBreakPoint, defaultTheme.smallBreakPoint);
-    updateStyle('$mediumBreakPoint', config.mediumBreakPoint, theme.mediumBreakPoint, defaultTheme.mediumBreakPoint);
+    updateStyle(styleTemplate, '$positionLeft', config.positionLeft, theme.positionLeft, defaultTheme.positionLeft);
+    updateStyle(styleTemplate, '$smallBreakPoint', config.smallBreakPoint, theme.smallBreakPoint, defaultTheme.smallBreakPoint);
+    updateStyle(styleTemplate, '$mediumBreakPoint', config.mediumBreakPoint, theme.mediumBreakPoint, defaultTheme.mediumBreakPoint);
 
-    updateStyle('$menuTextColor', config.menuTextColor, theme.menuTextColor, defaultTheme.menuTextColor);
-    updateStyle('$menuBackgroundColor', config.menuBackgroundColor, theme.menuBackgroundColor, defaultTheme.menuBackgroundColor);
+    updateStyle(styleTemplate, '$menuTextColor', config.menuTextColor, theme.menuTextColor, defaultTheme.menuTextColor);
+    updateStyle(styleTemplate, '$menuBackgroundColor', config.menuBackgroundColor, theme.menuBackgroundColor, defaultTheme.menuBackgroundColor);
 
-    updateStyle('$menuitemFocusTextColor', config.menuitemFocusTextColor, theme.menuitemFocusTextColor, defaultTheme.menuitemFocusTextColor);
-    updateStyle('$menuitemFocusBackgroundColor', config.menuitemFocusBackgroundColor, theme.menuitemFocusBackgroundColor, defaultTheme.menuitemFocusBackgroundColor);
+    updateStyle(styleTemplate, '$menuitemFocusTextColor', config.menuitemFocusTextColor, theme.menuitemFocusTextColor, defaultTheme.menuitemFocusTextColor);
+    updateStyle(styleTemplate, '$menuitemFocusBackgroundColor', config.menuitemFocusBackgroundColor, theme.menuitemFocusBackgroundColor, defaultTheme.menuitemFocusBackgroundColor);
 
-    updateStyle('$focusBorderColor', config.focusBorderColor, theme.focusBorderColor, defaultTheme.focusBorderColor);
+    updateStyle(styleTemplate, '$focusBorderColor', config.focusBorderColor, theme.focusBorderColor, defaultTheme.focusBorderColor);
 
-    updateStyle('$buttonTextColor', config.buttonTextColor, theme.buttonTextColor, defaultTheme.buttonTextColor);
-    updateStyle('$buttonBackgroundColor', config.buttonBackgroundColor, theme.buttonBackgroundColor, defaultTheme.buttonBackgroundColor);
+    updateStyle(styleTemplate, '$buttonTextColor', config.buttonTextColor, theme.buttonTextColor, defaultTheme.buttonTextColor);
+    updateStyle(styleTemplate, '$buttonBackgroundColor', config.buttonBackgroundColor, theme.buttonBackgroundColor, defaultTheme.buttonBackgroundColor);
 
-    updateStyle('$zIndex', config.zIndex, theme.zIndex, defaultTheme.zIndex);
+    updateStyle(styleTemplate, '$zIndex', config.zIndex, theme.zIndex, defaultTheme.zIndex);
 
-    updateStyle('$zHighlight', config.zHighlight, theme.zHighlight, defaultTheme.zHighlight);
+    updateStyle(styleTemplate, '$zHighlight', config.zHighlight, theme.zHighlight, defaultTheme.zHighlight);
+
+    updateStyle(highlightTemplate, '$buttonBackgroundColor', config.buttonBackgroundColor, theme.buttonBackgroundColor, defaultTheme.buttonBackgroundColor);
+    updateStyle(highlightTemplate, '$focusBorderColor', config.focusBorderColor, theme.focusBorderColor, defaultTheme.focusBorderColor);
 
   }
 
@@ -693,16 +700,24 @@ $skipToId-highlight div {
    *
    *   @desc  Updates the style sheet template and then attaches it to the document
    *
-   * @param {Object}  attachNode  - DOM element node to attach button and menu container element
+   * @param {Object}  attachNode      - DOM element node to attach button and menu container element
    * @param  {Object}  config          -  Configuration information object
    * @param  {String}  skipYToStyleId  -  Id used for the skipto container element
    */
   function renderStyleElement (attachNode, config, skipToId) {
     styleTemplate.innerHTML = styleTemplate.innerHTML.replaceAll('$skipToId', '#' + skipToId);
+    highlightTemplate.innerHTML = highlightTemplate.innerHTML.replaceAll('$skipToId', '#' + skipToId);
     addCSSColors(config);
+
     const styleNode = styleTemplate.content.cloneNode(true);
     styleNode.id = `${skipToId}-style`;
     attachNode.appendChild(styleNode);
+
+    const highlightNode = highlightTemplate.content.cloneNode(true);
+    highlightNode.id = `${skipToId}-highlight`;
+    const headNode = document.querySelector('head');
+    headNode.appendChild(highlightNode);
+
   }
 
   /* utils.js */
@@ -1603,7 +1618,7 @@ $skipToId-highlight div {
     let landmarkInfo = [];
     let targetLandmarks = getLandmarkTargets(landmarkTargets.toLowerCase());
     let targetHeadings  = getHeadingTargets(headingTargets.toLowerCase());
-    let onlyInMain = headingTargets.includes('main');
+    let onlyInMain = headingTargets.includes('main') || headingTargets.includes('main-only');
 
     function transverseDOM(startingNode, doc, parentDoc=null, inMain = false) {
       for (let node = startingNode.firstChild; node !== null; node = node.nextSibling ) {
@@ -1720,7 +1735,7 @@ $skipToId-highlight div {
     // If targets undefined, use default settings
     if (typeof headingTargets !== 'string') {
       console.warn(`[skipto.js]: Error in heading configuration`);
-      headingTargets = 'h1 h2';
+      headingTargets = 'main-only h1 h2';
     }
 
     const [landmarks, headings] = queryDOMForLandmarksAndHeadings(landmarkTargets, headingTargets, skiptoId);
@@ -3156,7 +3171,7 @@ $skipToId-highlight div {
 
         // Selectors for landmark and headings sections
         landmarks: 'main search navigation complementary',
-        headings: 'main h1 h2',
+        headings: 'main-only h1 h2',
 
         // Highlight options
         highlightTarget: 'enabled', // options: 'enabled' (default) and 'disabled'